pyproject.toml
[project]
name = "moodlehack"
version = "0.2.1rc1"
description = "Moodle test answers repo with fast search and REST API exchange"
readme = "README.md"
authors = [
    { name = "MOIS3Y", email = "stepan@zhukovsky.me" }
]
license = { text = "GPLv3" }
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.29.0",
    "django-filter>=25.1",
    "django-crispy-forms>=2.5",
    "django-typer>=3.3.2",
    "crispy-bootstrap5>=2025.6",
    "markdown>=3.10",
    "pydantic-settings>=2.10.1",
    "platformdirs>=4.5.0",
    "rich>=14.2.0",
    "starlette>=0.47.2",
    "uvicorn>=0.38.0",
]

[project.scripts]
moodlehack = "moodlehack.manage:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
